β0.1

+++ Assets/Scripts/Enemy.cs
@@ -8,12 +8,11 @@ public class Enemy : MonoBehaviour
     public int rotspeed;
     public int power = 1;
 
-    public int hp = 1;
+    public float hp = 1;
 
 
 
     Bullet playerBullet;
-
     PlayerShield playerShield;
 
     // Start is called before the first frame update
